==> 00_00_Definir_Rutas.R <==
# Definir rutas

# path : Ruta hacia la carpeta de trabajo
# pathData : Ruta hacia la carpeta de Datasets (contenida en la carpeta de trabajo)

path <- "C:/Users/andre/Desktop/4t/Mineria/Treball/"
pathData <- paste0(path,"Datasets/")
